mcp/requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
Authlib==1.6.0
beautifulsoup4==4.13.4
boto3==1.39.9
botocore==1.39.9
bs4==0.0.2
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cohere==5.16.1
colorama==0.4.6
cryptography==45.0.5
cyclopts==3.22.2
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastavro==1.11.1
fastmcp==2.10.5
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.26.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
groq==0.30.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
linkify-it-py==2.0.3
logfire==3.25.0
logfire-api==3.25.0
markdown-it-py==3.0.0
mcp==1.12.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistralai==1.9.2
multidict==6.6.3
openai==1.97.0
openapi-pydantic==0.5.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
packaging==25.0
platformdirs==4.3.8
playwright==1.53.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-ai==0.4.4
pydantic-ai-slim==0.4.4
pydantic-evals==0.4.4
pydantic-graph==0.4.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rich-rst==1.3.1
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.4.1
starlette==0.47.1
tenacity==8.5.0
tokenizers==0.21.2
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
uc-micro-py==1.0.3
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
